fix: call SIGINT to ensure vLLM child process cleanup
When vLLM is run with --tensor-parallel-size N,
where N > 1, there are child process created for
each of the N GPUs the model is sharded over.

When vLLM is shut down with SIGTERM these child
processes are not terminated. Sending a SIGINT
to the vLLM parent process does terminate the
child processes.

Signed-off-by: Ali Maredia <amaredia@redhat.com>
Co-authored-by: James Kunstle <jkunstle@redhat.com>