PHASE 1.2: Hardware & System Profile Autodetection
Hardware auto detection of profiles that wold apply
on top of the default configuration.

All work is in this commit is focused on the ilab model
chat & serve command.

This commit adds the following:
- Add 2 generic hardware profiles and Nvidia l4x4 profile
- Add hardware based profile auto detection and application
to the default config
- Enable click flags to read from the most up to
date configuration for default values and for
help text
- Enable model paths in configuration files to just be
the path under the model dir

Future work:
- Unit tests and functional tests
- Adding more profiles and general hardware autodetection
- UX refinement of output messages, logging,
and errror messages
- determine how autodetection and manual interactive
profile selection will interact

Signed-off-by: Ali Maredia <amaredia@redhat.com>