eval: profile changes for updated eval config changes
This commit does the following:

1. The output dir for mt_bench used to be:
~/.local/share/instructlab/internal/eval_data

It is now:
~/.local/share/instructlab/internal/eval_data/mt_bench

This commit explicitly changes that in all of the
profiles though this default shoud apply on it's own.

2. Overrides the judge model in the
config section for mt-bench-branch to match
the judge model used in mt-bench for smaller hardware
profiles.

3. Updates the evaluation section of the config for the
custom profiles for CI. Though many of the values added
are not changed from defaults they are added as a reference
in case further changes want to be made in the future.

Signed-off-by: Ali Maredia <amaredia@redhat.com>